chore: move `cast send` tests
